backend/requirements.txt
# Core Libraries
numpy==1.24.3
pandas==1.5.3
scipy==1.11.3
scikit-learn==1.7.2

# OCR and Image Processing
pillow==10.0.1
pytesseract==0.3.10
pdf2image==1.16.3

# Web Frameworks
fastapi==0.119.0
uvicorn==0.23.2

# Flask (Remove if not used)
Flask==3.1.2
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0

# Pydantic and Starlette
pydantic==2.12.0
pydantic_core==2.41.1
starlette==0.48.0

# Utilities
joblib==1.5.2
threadpoolctl==3.6.0
python-multipart==0.0.20
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
typing-inspection==0.4.2
typing_extensions==4.15.0
anyio==4.11.0
h11==0.16.0
idna==3.10
six==1.17.0
sniffio==1.3.1
python-dotenv==1.0.0